S3-T14: CI release gate with DB guard, full suites, and Playwright smoke
DB-backed tests import a shared guard that hard-fails unless DATABASE_URL
is localhost/clipmind_dev or CLIPMIND_TEST_DB_OVERRIDE=1 (CI service
containers only). Push tests scope to fixture creators via a creatorIds
test filter instead of disabling every other creator's push. CI now runs
six gating jobs with no continue-on-error: env-file hygiene, full
typecheck, all pure Node suites (108), DB suites (85) against postgres:17
with migrate deploy, the Playwright smoke (build + seed + login + Home/
Review/Rhythm render), and clip-service ruff + pytest. E2E deliberately
excludes upload/clip-service/Minds/OpenAI paths - documented in README.

Verification fixes: strict-mode locator in the smoke (2 seeded clips),
and 10 ruff findings in clip-service (RUF046 casts, import sort, G201
logger.exception) so the new ruff gate starts green.

Local proof: test:db 85/85, test:node 108/108, test:e2e 1/1 passed,
pytest 43/43, ruff all checks passed, typecheck clean.

+++ clip-service/src/candidates.py
@@ -311,7 +311,7 @@ def loudness_samples(video_path: Path) -> list[LoudnessSample]:
 
     samples: list[LoudnessSample] = []
     for match in LOUDNESS_RE.finditer(result.stderr):
-        time_ms = int(round(float(match.group("time")) * 1000))
+        time_ms = round(float(match.group("time")) * 1000)
         momentary = float(match.group("momentary"))
         samples.append(LoudnessSample(time_ms=time_ms, momentary_lufs=momentary))
